+++ LoveBird/Sources/Depedency/DependencyValues+Extension.swift
@@ -9,12 +9,32 @@ import ComposableArchitecture
 import Moya
 
 extension DependencyValues {
+  var userData: UserData {
+    get { self[UserData.self] }
+    set { self[UserData.self] = newValue }
+  }
+  
   var lovebirdApi: LovebirdAPI {
     get { self[LovebirdAPI.self] }
     set { self[LovebirdAPI.self] = newValue }
   }
+
+  var appConfiguration: AppConfiguration {
+    get { self[AppConfiguration.self] }
+    set { self[AppConfiguration.self] = newValue }
+  }
+}
+
+// MARK: - `DependencyKey` Implementation
+
+extension UserData: DependencyKey {
+  static public let liveValue = Self()
 }
 
 extension LovebirdAPI: DependencyKey {
   public static var liveValue = LovebirdAPI(apiClient: MoyaProvider<APIClient>())
+}
+
+extension AppConfiguration: DependencyKey {
+  static public let liveValue = AppConfiguration()
 }
